refacto(provider): introduce ProviderCapabilitiesRepository instead of shared module in use-cases
