+++ api/use_cases/admin/providers/_createproviderusecase.py
@@ -1,12 +1,11 @@
 from dataclasses import dataclass
 
 from api.domain.model.errors import InconsistentModelMaxContextLengthError, InconsistentModelVectorSizeError
-from api.domain.provider import ProviderAdapterBuilder, ProviderClient, ProviderRepository
+from api.domain.provider import ProviderCapabilitiesRepository, ProviderRepository
 from api.domain.provider.entities import BasicAuth, HostingZone, Metric, Provider, ProviderType
 from api.domain.provider.errors import InvalidProviderTypeError, ProviderAlreadyExistsError, ProviderNotReachableError
 from api.domain.router import RouterRepository
 from api.domain.router.errors import RouterNotFoundError
-from api.use_cases.provider import get_provider_capabilities
 
 
 @dataclass
@@ -43,11 +42,10 @@ type CreateProviderUseCaseResult = (
 
 
 class CreateProviderUseCase:
-    def __init__(self, router_repository: RouterRepository, provider_repository: ProviderRepository, provider_client: ProviderClient, provider_adapter_builder: ProviderAdapterBuilder):
+    def __init__(self, router_repository: RouterRepository, provider_repository: ProviderRepository, provider_capabilities_repository: ProviderCapabilitiesRepository):
         self.router_repository = router_repository
         self.provider_repository = provider_repository
-        self.provider_client = provider_client
-        self.provider_adapter_builder = provider_adapter_builder
+        self.provider_capabilities_repository = provider_capabilities_repository
 
     async def execute(self, command: CreateProviderCommand) -> CreateProviderUseCaseResult:
         router = await self.router_repository.get_router_by_id(router_id=command.router_id)
@@ -57,9 +55,7 @@ class CreateProviderUseCase:
         if not command.provider_type.is_compatible_with(router_type=router.type):
             return InvalidProviderTypeError(provider_type=command.provider_type.value, router_type=router.type.value)
 
-        result = await get_provider_capabilities(
-            provider_client=self.provider_client,
-            provider_adapter_builder=self.provider_adapter_builder,
+        result = await self.provider_capabilities_repository.get_provider_capabilities(
             router_type=router.type,
             provider_type=command.provider_type,
             url=command.url,
