--- api/use_cases/provider/__init__.py
@@ -1,5 +1,0 @@
-from ._getprovidercapabilities import get_provider_capabilities
-
-__all__ = [
-    "get_provider_capabilities",
-]
